% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count.R
\name{amino_frequency_plot_local}
\alias{amino_frequency_plot_local}
\title{Plot frequency of amino acids}
\usage{
amino_frequency_plot_local(input_string)
}
\arguments{
\item{input_string}{A character string to analyze for unique character frequencies.}
}
\value{
A ggplot object displaying a bar plot of unique character counts.
}
\description{
Plot frequency of amino acids
}
\examples{
# Example usage
plot <- amino_frequency_plot_local("AABBCDAA")
print(plot)  # Print the plot
}
