% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dna.R
\name{clean_dna}
\alias{clean_dna}
\title{Clean DNA
clean_dna function removes "N" nucleotides from raw DNA sequence in FASTA format and deletes new lines.}
\usage{
clean_dna(raw_seq)
}
\arguments{
\item{raw_seq}{Raw sequence of DNA in fasta format.}
}
\value{
dna_seq
}
\description{
Clean DNA
clean_dna function removes "N" nucleotides from raw DNA sequence in FASTA format and deletes new lines.
}
\examples{
No examples available.

}
