% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codonMerging.R
\name{codonMerging}
\alias{codonMerging}
\title{Creation of mRNA template for translation}
\usage{
codonMerging(codons)
}
\arguments{
\item{codons}{codon list to be merged}
}
\value{
mRNA template as a single string sequence
}
\description{
Creation of mRNA template for translation
}
\examples{
codons <- c("AUG", "UCG", "UUU")
codonMerging(codons)

}
