% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codons}
\alias{codons}
\title{Codons vector}
\format{
An object of class \code{character} of length 64.
}
\usage{
codons
}
\description{
Description
64 codons and their corresponding amino acids on letter code. i is a in vector form.
}
\keyword{datasets}
