% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finds_codons.R
\name{finds_codons}
\alias{finds_codons}
\title{finds_codons}
\usage{
finds_codons(sequence, start = 1)
}
\arguments{
\item{sequence:}{RNA sequence}

\item{start:}{starting codon}
}
\value{

}
\description{
finds_codons
}
\examples{
function('AGCCUAUGUACUGUAGUCGUAGUCUAA')
}
