% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_synthesis.R
\name{prot_synthesis}
\alias{prot_synthesis}
\title{prot_synthesis}
\usage{
prot_synthesis(codon_list)
}
\arguments{
\item{codon_list}{}
}
\value{
prot
}
\description{
prot_synthesis
}
\examples{
prot_synthesis(codons)
}
