% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_dna.R
\name{random_dna}
\alias{random_dna}
\title{Create random DNA string of a set lenght}
\usage{
random_dna(lenght_of_dna)
}
\arguments{
\item{lenght_of_dna}{setting the lenght of the random DNA string you want}
}
\value{
A random dna string of length lenght_of_dna randomly chosen from the nucleic bases.
}
\description{
Create random DNA string of a set lenght
}
\examples{
random_dna(lenght_of_dna = 10)

}
