% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacement_of_matches.R
\name{replace_T_with_U}
\alias{replace_T_with_U}
\title{Replace T with U}
\usage{
replace_T_with_U(DNA_sequence)
}
\arguments{
\item{DNA_sequence}{}
}
\value{
RNA sequence
}
\description{
Replace T with U
}
\examples{
replace_T_with_U(DNA_sequence = "GATTA")
}
