% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaProcessing.R
\name{rnaProcessing}
\alias{rnaProcessing}
\title{RNA processing step}
\usage{
rnaProcessing(rnaSequence, start = 1)
}
\arguments{
\item{rnaSequence}{sequence of rna nucleotides}

\item{start}{define the start position on the string}
}
\value{
Processed RNA sequence
}
\description{
RNA processing step
}
\examples{
# Testing a random RNA sequence
rnaSequence <- "AUGCAUGCAUGCU"
rnaProcessing(rnaSequence)
}
