% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_to_codons.R
\name{seq_to_codons}
\alias{seq_to_codons}
\title{seq_to_codons}
\usage{
seq_to_codons(sequence, start = 1)
}
\arguments{
\item{sequence}{A nucleotide sequence}

\item{start}{The first position in the nucleotide sequence that should start the codon splitting.}
}
\value{
The codons of the input sequence is outputted as a vector of strings.
}
\description{
This is a function that takes a nucleotide sequence and splits it into the codons, starting from position indicated in start
}
\examples{
seq_to_codons("ACTGGTCAGATCATGA", start=1)
}
