% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceGeneration.R
\name{sequenceGeneration}
\alias{sequenceGeneration}
\title{New random sequence creation}
\usage{
sequenceGeneration(sequenceLength)
}
\arguments{
\item{sequenceLength}{desired sequence length}
}
\value{
random DNA sequence
}
\description{
New random sequence creation
}
