% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_codon_table}
\alias{standard_codon_table}
\title{standard_codon_table}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
standard_codon_table
}
\description{
translations of codons to amino acids
}
\keyword{datasets}
