% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcription.R
\name{transcription}
\alias{transcription}
\title{DNA to RNA transcription}
\usage{
transcription(dnaSequence)
}
\arguments{
\item{dnaSequence}{String composed of singleletter nucleotide code of variable sequence.}
}
\value{
A new sequence with all the thymines substituted by uracil.
}
\description{
DNA to RNA transcription
}
