% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codons.R
\name{translate_codons}
\alias{translate_codons}
\title{translate_codons}
\usage{
translate_codons(codons)
}
\arguments{
\item{codons}{Vector containing triplet(s) of nucleotides}
}
\value{
String of the translated amino acid sequence without spacing
}
\description{
Translates codon(s) of nucleotides to a string of amino acids (AAs)
}
\examples{
codons <- c("AUG", "UUU", "UGA")
translated_sequence <- translate_codons(codons)
}
