% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation form RNA to protien sequence}
\usage{
translation(mRnaTemplate)
}
\arguments{
\item{mRnaTemplate}{RNA sequence input}
}
\value{
plot with aa distribution of the sequence
}
\description{
Translation form RNA to protien sequence
}
\examples{
mRnaTemplate <- "AUGUCGUUU"
translation(mRnaTemplate)

}
