% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/function_1.R
\name{clean_dna}
\alias{clean_dna}
\title{Clean DNA
clean_dna functino removes "N" nucleotides from raw DNA sequence in FASTA format and deletes new lines.}
=======
% Please edit documentation in R/clean_dna.R
\name{clean_dna}
\alias{clean_dna}
\title{Clean DNA
clean_dna function removes "N" nucleotides from raw DNA sequence in FASTA format and deletes new lines.}
>>>>>>> bf6f2cd9d07c137163ae09ec37a6de932a91e1bd
\usage{
clean_dna(raw_seq)
}
\arguments{
\item{raw_seq}{Raw sequence of DNA in fasta format.}
}
\value{
dna_seq
}
\description{
Clean DNA
<<<<<<< HEAD
clean_dna functino removes "N" nucleotides from raw DNA sequence in FASTA format and deletes new lines.
=======
clean_dna function removes "N" nucleotides from raw DNA sequence in FASTA format and deletes new lines.
>>>>>>> bf6f2cd9d07c137163ae09ec37a6de932a91e1bd
}
\examples{
No examples available.

}
